Enable logs for frontend fargate instance
This makes a couple of minor environment variable fixes for
the frontend application to run in Fargate (for now we are
pointing Plek/Slimmer/frontend to production for resources
it needs to run).

This also adds logging to the frontend service, so one can
view the application logs for frontend instances in a single
place - either in fargate's service logs or in CloudWatch.