Add Statsd project
This will run statsd in a container instance in AWS ECS Fargate.

Currently statsd runs on all EC2 instances. Since we are doing a
discovery into moving our applications to ECS Fargate, running
statsd in all containers seems a bad idea. Instead we intend to
run one cluster/service/task of statsd, and have each container
talk to that statsd.

Currently this uses the statsd/statsd image provided by the statsd
project. This uses the latest master of that project.
Note that statsd is not configurable with command line arguments,
and the image must be built with a config file. Therefore we should
in future build our own image, as the sensible defaults provided
by statsd in their official image released to DockerHub are not
quite what we use in production (see govuk-puppet statsd config).

This should enable us to test whether our new container instances
living in Fargate will be able to talk to Graphite living in EC2.
If this works, it will enable us to delay thinking about app
metrics until it comes time to remove the Graphite machine in EC2.

This change also splits out the security groups necessary for
ECS tasks into the fargate-security project. In a new environment
this project will need to be built first.