Add GOVUK_APP_DOMAIN_EXTERNAL envvar to Publisher service
This environment variable is used for external URLs, e.g. when
redirecting users to the public Signon URL.

It's essential since Plek falls back to GOVUK_APP_DOMAIN when
GOVUK_APP_DOMAIN_EXTERNAL is not set. This routes users
to signon.<env>.govuk-internal.digital which of course
is inacessible to end users.