Add frontend-ecs.test.govuk-internal.digital DNS record
This record will allow us to route traffic to the frontend machine
on frontend.pink.test (including switching back to the frontend app
on that machine), and also route traffic to the frontend application
running in ECS at frontend-ecs.test.

I propose that we use DNS changes where possible to route traffic to
services in ECS. However, Frontend is a special case, since frontend
is also used for both the frontend machine and application.

To route traffic to frontend in ECS, I've created this new DNS record
and updated the Frontend 'Backend' in the test account's deployment of
Router API to send requests for frontend to this new host.