Create GOV.UK Cluster
The intent is to have a single module called govuk as the root
of all other modules, so that a single terraform apply command
can bring up the entirety of the GOV.UK infrastructure.

This module will create a single cluster for govuk under
which all services will run (publishing apps, frontend apps).