Use AppMesh for internal networking in ECS
We have an opportunity to change how we do internal networking with the move
from EC2 to ECS.

We also know that applications in EC2 will not send requests to services in
ECS. Applications in ECS will still send requests out to external services
and services in EC2 (namely, self-hosted databases we have not yet migrated).

We have decided to adopt a service mesh for internal networking for services
in ECS. We will use AWS Cloud Map for service discovery within an envoy-based
service mesh managed by AWS App Mesh.

Ingress into the mesh could be managed by Virtual Gateways (part of App Mesh).
However, we are choosing to use ALBs for ingress.

In addition, this change demonstrates our intent to experiment with a
monolithic terraform module (a single module one calls `terraform apply` in),
which includes all other modules required to bring up the GOV.UK infrastructure.
This has some downsides (longer apply times, additional complexity of applies,
too-close coupling of services) but it brings operational simplicity in that
one only needs to run `terraform destroy` or `terraform apply` to make changes
to the entire new infrastructure, rather than requiring us to write imperative
wrapper scripts that modify the infrastructure in a particular order. It will
also (we hope) make graphing and managing dependencies between dependencies
a bit easier. If it doesn't work we will refactor this into smaller modules.

This starts by adding Publisher and Publishing API to App Mesh.

Co-authored-by: Richard Towers <richard.towers@digital.cabinet-office.gov.uk>
Co-authored-by: Mahmud Hussain <mahmud.hussain@digital.cabinet-office.gov.uk>