Enable Publisher to talk to Publishing API via App Mesh
This changes the security groups of publishing-api and publisher, to
enable publisher to make requests to publishing api.

This also changed the service discovery namespace to a
'aws_service_discovery_private_dns_namespace' which is probably not what
we want - I'll see if we can change this in a followup commit.

The 'user' property of the envoy container definition (1337) is actually
required and must match the IgnoredUID set on the ECS service
proxy_configuration. The proxy_configuration has many magic numbers,
such as 15001 and 15000 - change them at your peril.

Additionally a task_role is now required by task defitions in the mesh
to give the Envoy services permission to access the App Mesh APIs to
read their configuration using the AWS managed AWSAppMeshEnvoyAccess IAM
policy.

I've left the security group 'sg-05ad7398fc0d7c5b4' attached to the
publishing-api service, since though the name 'govuk_publishing-api_access'
and inbound rules (permitting ELB access) suggest this is not necessary
now that publishing-api doesn't have a load balancer, this security group
is also used in inbound security group rules on the Redis and Postgres
AWS-managed service security groups to give publishing-api access to them.
So this security group is still necessary.

We should aim for all services that an ECS service depends on to have its
URL specified as an env var such as PLEK_SERVICE_PUBLISHING_API_URI to
gradually phase out any Plek involvement in service discovery/resolution.

Followup commits may switch back the namespace to 'API only' so we don't
rely on DNS. I'll also change govuk.local to internal.govuk since as
Phil P mentioned, local is reserved for multicast DNS and some resolvers
(including systemd-resolved) refuse to send requests to recursors.
Or I may call it mesh.govuk-internal.digital - depends if CloudMap permits
us to us a dot in the namespace if it's not using DNS.

Still to do: Bring content-store into the mesh.