Change Cloud Map namespace from API calls and DNS queries in VPCs to API Calls
You have a few options when creating a CloudMap namespace. From the AWS documentation:

- API calls: Your application can find registered instances by submitting DiscoverInstances
  requests to the AWS Cloud Map API.
- API calls and DNS queries in VPCs: Your application can find registered instances both by
  submitting DiscoverInstances requests to the AWS Cloud Map API and by submitting DNS queries
  in a VPC.

This commit changes the namespace to API calls only. The consequence of this is that applications
no longer are started with a fixed URL for a service, e.g. a PLEK_SERVICE_CONTENT_STORE_URI
environment variable. This is because there will no longer be DNS records for services: only
IP addresses.

One downside of this is that requests may be slower. We may need to poll and cache the result
of DiscoverInstances CloudMap requests in applications. If the router application made a
request to the AWS CloudMap API for every request, this would add additional latency to end
user requests.

Another downside of this approach is that we must instrument applications to discover services,
using an AWS SDK for service discovery. This requires applications to know more about infrastructure
than we would like.