Make frontend a non-root module like the other apps.
Hook frontend into the modules/govuk composition module and make it
look like the others, copying the same stuff for using AppMesh and so
on.

Still contains a ton of duplication because we haven't factored out the
common ECS task/service stuff into a module yet, but this at least gives
us a consistent base from which to modularise.