WIP - move container_definitions inline
We can use jsonencode to define container_definitions without having to
load them from a different file.

This would allow us to use the full "power" of HCL to template out the
JSON, which could help us avoid repeating huge chunks of it.

It would probably also allow us to grab the references to secrets from
secrets manager based on their friendly name, instead of hardcoding the
ARNs like we're doing now (
https://registry.terraform.io/providers/hashicorp/aws/latest/docs/data-sources/secretsmanager_secret
)

Mostly just pushing this commit as a proof of concept - it would require
a bit more refactoring to make this nice.