Get secretsmanager ARNs using terraform data
This means we don't have to hardcode things which are obviously dynamic
/ random.

Doing things this way is still not entirely ideal though, because:

- Getting secrets by name like this means they're global to an AWS
  account - what if we want to run more than one copy of the platform
  with different secrets in the same AWS account
- The dependency between the terraform and which secrets need to be in
  secretsmanager isn't clear - at the moment these secrets were added by
  hand, but it would be nice to try to automate that as much as
  possible...

Some of these OAUTH_ID / OAUTH_SECRET things look like they were
probably created by signon, which is not going to be a great pattern.

Ideally we'd be able to generate all the secrets the apps need before
running the apps themselves (otherwise there's an annyoing bootstrapping
issue).