Factor out common app resources into an `app` module.
Apply this just to content-store for now, others to follow the same
pattern shortly.

Refactor and simplify the way we're defining the security groups, in
order to achieve a stable interface to the `app` module. We only need
one security group per app (service) and we don't want to have to modify
the interface of the per-app modules (like content-store and
publishing-api) whenever we want to change a firewall rule. Therefore
move the security group rule definitions into the `govuk` composition
module.

Clean up a few descriptions and names of things along the way and
sprinkle some TODOs.