Clean up ALB config for Publisher and factor out domain name.
This introduces a new top-level var public_lb_domain_name which looks a
bit like a duplicate of govuk_app_domain_external as it's set to the
same thing ("test.govuk.digital") but they're for quite different
purposes.

I think it's likely that we'll need to control them separately because
govuk_app_domain_external will at some stage need to be the
$env.publishing.service.gov.uk domain whereas Terraform will be creating
aliases for load balancers in $env.govuk.digital (with some
externally-managed CNAMEs linking the two together). We don't know how
the details of this will play out in the transition to ECS yet.

The name public_lb_domain_name probably isn't the best, but hopefully at
least it makes it clear what it's for at the moment.

Also:
 - Enable evaluate_target_health.
 - Remove a redundant domain suffix from a Route53 resource (this was a
   common antipattern in govuk-aws so was probably copied from there).
 - Improve some descriptions.