Attach Publisher's load balancer to its ECS service.
Add a parameter to `modules/app` to allow attaching ALB/NLB target
groups to an app's ECS service. Unfortunately we have to use Terraform's
rather clunky "dynamic" construct in order to do this because the AWS
provider doesn't have a separate resource type for associating an ECS
service with a load balancer (unlike for EC2, where it provides
aws_lb_target_group_attachment).

Define the healthcheck grace period on the ECS service if we're
attaching one or more load balancers. Parameterise this as well, as
we're likely to need to vary it. We have to avoid specifying
health_check_grace_period_seconds on the aws_ecs_service resource if
there are no load balancers, because otherwise the AWS TF provider
complains; hence the ugly ternary conditional there.