Add draft-content-store
Following on from #13 and #22 which added content-store, we add the
draft version of content-store in order to build the draft stack of
GOV.UK.

We customize the existing content-store app module so that the
service_name variable can be used to differentiate between draft and non
draft app deployment.

Secrets are renamed and will have to be changed via CLI

Future work:
1. SENTRY_DSN secret is the same across all apps in same environment so
this needs to be fixed throughout the repo.

Ref:
1. [Task Trello Card]
(https://trello.com/c/AaBscxsV/262-spin-up-draft-content-store)
2. [Parent Trello Card]
(https://trello.com/c/X4jlkXyE/200-determine-a-minimal-subset-of-apps)