Refactor Publishing API to use task definition module
This factors out the configuration for task definitions from the app
module to the task definition module.

Publishing API will have the same configuration, so this should be a
no-op in AWS.

The default deployed-to-<env> image tag is used to bring up the task
definition for a service. A lifecycle rule on the ECS service ensures
that this only occurs during the bootstrapping phase. After an initial
terraform apply, the app module will no longer be responsible for
creating new task definition revisions. This permits a deployment
module to create new task definitions and store the state in a separate
statefile, permitting concurrent and independent deploys.

Further commits will make the other apps consistent with this change.

Co-authored-by: Stephen Ford <stephen.ford@digital.cabinet-office.gov.uk>