Use tfvar files to share configuration between apps and infra
This introduces a new directory: `terraform/deployments/variables`
which will contain the variable definitions (.tfvar) files that should
be called when applying terraform for an app task definition or for
the infrastructure.

The new workflows:

- Applying the infrastructure
  ```shell
  $ cd terraform/deployments/govuk-test
  $ terraform apply -var-file=../variables/test/common.tfvars -var-file=../variables/test/infrastructure.tfvars
  ```shell
- Applying the application
  ```
  $ cd terraform/deployments/apps/frontend
  $ terraform apply -var-file=../../variables/test/common.tfvars -var-file=../../variables/test/apps.tfvars -var=image_tag=release_123
  ```

The variables/test/common.tfvars file contains variables common to the test
environment deployments.

Should we want to create a new environment we would need to duplicate these
tfvars files for another environment. When we add extra environments we may
find that we can introduce a cross-environment common.tfvars file.