Pass assume_role_arn to AWS terraform provider
This allows users who need to assume a role to run terraform to pass the
ARN of the role to terraform. Terraform can assume roles as part of the
provider configuration.

Users who don't need to assume a role can simply not pass the variable.
It defaults to `null`, in which case terraform will just use whatever
AWS credentials it finds in the environment.