Grant concourse additional IAM permissions for TF
To be able to run the app terraform, concourse needs to have permission
to:

- read and write the statefile in S3
- describe secrets in SecretsManager to look up their IDs
- describe IAM roles to look up their IDs
- administer ECS (which it already could)

I've put some TODO comments in the config because I used pre-defined AWS
policies which are more permissive than we need.

There probably needs to be a discussion about how much permission
we give concourse. If it's going to run the infrastructure terraform,
it's going to need a lot of permission. Hand crafting a restrictive set
of IAM policies that allow it to do everything it needs to, but nothing
more would be a lot of work. While we're actively developing, it might
be easier to give it a poweruser role.

On the other hand, concourse really doesn't need much permission to
deploy the apps, so maybe we could hand craft a more-restrictive role
for these tasks. Concourse could assume the highly privileged role to do
infrastructure, and the less privileged role to deploy apps. Not sure if
there's any practical security benefit to that though.