Split deploy-app into task-def / ecs-service tasks
This allows us to use a terraform container in the task that does
terraform things, and an aws container in the task that does AWS things.

It also means we're doing slightly less orchestration in shell scripts,
and slightly more in concourse which is probably a good thing.

If we wanted to we could probably use a concourse terraform resource
instead of the update-task-definition task (it basically just runs
terraform apply). It doesn't look like there's an existing resource type
that would do what we want the update-ecs-service task to do, and it's
probably not worth writing one.