Use bootstrap image during bootstrapping
We want to deploy apps outside of Terraform, so don't want to provide
or manage task definitions in the govuk module.

However, the aws_ecs_service requires a task_definition to be set, in
order to permit service_registries and load_balancer being set.

Our hack solution we're trying out here is to use a dummy image containing
the envoy proxy during bootstrapping only. We'll then bring up the apps
as normal from the app deployment modules.

One downside is that we'll live in fear of someone removing the
ignore_changes = [task_definition] line from the ECS service. It would be
an extremely small change that could bring the system down.