Output govuk publisher_security_groups and private_subnets
These values are required as part of the network configuration to run tasks
for the Publisher Task Definition using ECS RunTask.

A subsequent commit will use these outputs to run one-off tasks in ECS.

We'll need to do the same for every service that we want to run tasks
for.