[spike] Branch deploys to ECS Fargate
This spike demonstrates one way to enable developers to deploy
branches to a production-like environment (integration).

Prequisites:
* We remove integration from the production deployment pipeline
* Releases are automatically deployed to the staging environment
* We deploy the latest release of all apps to integration overnight

This change requires that we maintain an additional pipeline for
each app we wish to run in integration. It would likely contain
duplicated code from the main deploy.yml pipeline.

This enables developers to modify the pipeline using
fly set-pipeline, to deploy branches without impacting other
developers, and crucially without impacting or requiring
authorization to modify the main deployment pipeline.

Integration would become a shared testing environment which would
be reset to a clone of staging daily.

Functionally, the only difference to the current situation is that
app releases would not necessarily be deployed to integration.

This enables us to avoid complicating the production deployment
pipelines, and draws a line between a testing environment
(integration) and the production environments (staging, production).

Other ideas are being evaluated - this is just a spike for
demonstration purposes.

Story: https://trello.com/c/5EN61qZx/231

Other ideas we're looking at:

- improving the local development experience so that it is not
  necessary to deploy branches
- modifying the production deployment pipeline so that it supports
  branch deploys
- creating a separate AWS environment per developer