Add security group rule for Publishing API to access Postgres.
Publishing API needs to talk to the existing RDS Postgres database. Add
a data resource to look up the existing security group (which is managed
from the old govuk-aws repo) and add a rule to allow these connections.

Also tweak the descriptions of the existing variables for similar
security groups.