Add temporary tool for running rake tasks
This is a tool to enable running ad-hoc tasks. It's expected to live
a few months until GDS CLI provides this feature. Building this into
GDS CLI at the moment is probably premature since the project may
change significantly in the next few months, so this tool should meet
our needs for now.

The tool prints a warning to users to let them know it's not going to
be around for long.

Expected usages:

run-task.sh -a frontend rake my:task
run-task.sh -a publisher curl http://publishing-api.mesh.govuk-internal.digital

This allows one to perform temporary tasks easily.