Factor out ContainerDefinition into new module
This creates a new module for ECS container definitions.

This is useful, as it removes some boilerplate from the task
definitions (envoy config, log config), and makes it cleaner
to create multiple container definitions within a task
definition.

This also demonstrates how we can horizontally scale the number
of containers within a container definition.

However, we should be careful here. If we increase the number
worker_count we should also take care to increase the CPU and
Memory of the Task Definition, too, to prevent OOM errors.