Create Concourse tasks for Publisher Worker deploys
This allows us to deploy the Publisher and Publisher Worker ECS services
in a single job, helping to ensure that these services run the same version
of the application.

There's a bit of a trade off here that we deploy the ECS Service for the web
process first, but I think that's acceptable - workers need to be backwards
compatible anyway (since very old jobs can conceivably get picked up by a new
worker).