create draft stack
As part of the publishing end-to-end journey, there is a dependency for
the draft stack to be up and running. This PR sets up the following
draft apps:
1. frontend
2. static
3. router
4. router-api

Draft content-store was set up in #81

It will be future work to verify that the publishing workflow works.

We have tested that the terraform applies correctly and concourse can
deploy all the new apps.

The draft frontend journey works also.

Ref:
1. [task trello card](https://trello.com/c/80luobV1/298-spin-up-draft-router-draft-router-api-draft-frontend)
2. [epic trello card](https://trello.com/c/bVSrlSQM/170-epic-bring-up-a-representative-set-of-apps-in-ecs-fargate)