Rename publisher service and container to publisher-web
This should make it clearer what this service does, since its name
now maps to a process defined in the Procfile.

There's a bit more configuration around the mesh subdomain to allow
services to still send requests to publisher.mesh rather than
publisher-web. This may not be necessary but I think it'll save
confusion if the URL stays the same.