Add a simple AWS billing dashboard
This is possible now that we've signed in to the root account and
enabled IAM users to view billing information :tada:.

This will hopefully give us a slightly easier-to-access view of our
billing (compared to having to sign in to AWS and look in
cost-explorer).

It's a bit naïve - CloudWatch's EstimatedCharges metric just increases
throughout the month and then resets. Might be nice to try to measure
the gradient, so we can see the rate we're spending at. But this
is good enough for now.

One interesting thing the graphs show is that the cost of the `test`
environment is still dominated by EC2 and RDS, even though we have quite
a bit of ECS stuff in there now. Not very scientific, but still.

Tested locally with docker-compose.