Use Fargate spot instances in the test environment.
Add FARGATE_SPOT as an allowable capacity provider in our ECS cluster
and make the default provider configurable.

The default is deliberately kept as on-demand and only the test
environment specifies spot, so that prod will have the right setting by
default even if this gets forgotten.

Tasks can still override the default capacity provider, so this doesn't
affect our ability to do things like have one-off user tasks run as spot
instances if we want (though we're not making use of that anywhere yet).

Motivation for this:

- Saves money.
- Will help to surface any potential issues when instances go away (and
  it's critical for reliability that we find such issues).
- Unlikely to hinder development (and if it does, it's easy to switch
  off temporarily or even permanently).