Remove json-check.sh
This was added when it looked like we were going to have a lot of JSON
alongside our terraform. We'd got configuration for things like
container definitions in JSON files, which were pretty complex.

The way things are leaning now, it looks like the vast majority is
actually going to be HCL (using the jsonencode() function to convert it
to JSON when needed).

Removing this check simplifies things ever so slightly.