Install terraform with setup-terraform
Instead of `brew` and `tfenv`.

This should have much better caching than the brew / tfenv combination,
which should speed up our actions a fair bit.

We already have the terraform verion in multiple places (concourse,
tfenv, now github actions too). I don't think there's an elegant way to
have a single-source-of-truth for them, so I think the best bet is
probably to write a test to check that they're in sync.