Add rspec tests to check versions are in sync
We've got a few places popping up where versions of things will need to
be kept in sync. For example, we're using terraform 0.13.3 locally, and
in concourse, and in GitHub Actions. There's no simple way of making all
these tools read from a canonical source, so instead let's just add some
tests that things are in-sync.

I'm using ruby / rspec for these tests, because this is
govuk-infrastructure and govuk developers are familiar with ruby /
rspec. This adds a new set of versions to keep in sync 🙈.

I hope that these tests will expand to cover more areas (not just
version syncing) in future.

One thing to note is that my use of `.dig()` is a bit fragile - if we
move tasks around, these tests will break. I could do it in a more
robust way (i.e. by specifying "the task with name run-terraform"
instead of specifying "the task with index 1"), but it would be a bit
more code.