spec: dig values from YAML more reliably
Previously I was using Hash#dig to specify the particular bit of YAML
that we wanted to check. That meant using indexes to effectively say
"the second job in the pipeline". That approach isn't very stable - if
we add a new job to the beginning of the pipeline, the test will fail.

Instead, we can use Array#find to say "the job in the pipeline with the
name 'run-terraform'", which should be more stable.