Add worspace aware domains and certificate
Domains are added to be managed by terraform, in the case of a
non default workspace, new subdomains are created. These new
domains will be use for the LB and the CNAMES of the apps.
Wokspace aware certificate are also created.

At this stage it is not clear how domains and certificate will
be managed in relation to workspaces, so this work is not complete
, workspace certificate are only in place on public LB and
delegation of the zones is not applied.