Add Statsd environment variables to apps
Apps can configure the host and protocol of Statsd, and
the namespace for all metrics sent by the app via
environment variables set in this commit.

The namespace govuk.app.<app_name>.ecs will make it easier
to include ECS-sourced metrics in graphs and alerts in the
existing Grafana charts / Icinga alerts, since the format
for apps in EC2 is: govuk.app.<app_name>.<ip_address>,
which is used as (e.g.) govuk.app.<app_name>.*.<metric>
in Grafana.