Add egress rules to all ECS Services
When using the Amazon provided DNS in VPC, Fargate tasks need
have outbound traffic allowed for HTTPS. This is in order to
pull the Envoy image from the ECR repository.

Apps also need to communicate with third-party APIs over the
Internet (Sentry, Notify, Google, etc).

Apps also need to send HTTP requests to other apps in the mesh.

This seemed like a better approach than adding lots of fine grained
rules immediately, or almost no security group rules. We plan
to revisit this approach once we know a bit more, since we're
still in a prototyping phase of the project.