Allow ingress from statsd in ECS to graphite in EC2
This adds the govuk_management_access security group
to the statsd ECS service

Graphite (hosted in EC2) has two security groups
• govuk_graphite_access for ELB ingress
• govuk_management_access for ingress from Statsd

Of course, govuk_management_access does more than that,
it also makes it possible to deploy to the service.
So Statsd could SSH onto the graphite instance if it
wanted to. We plan to remove govuk_management_access
from govuk-infrastructure in future, as it is poorly
named and is a bit of a bag of rules.