add monitoring concourse pipeline
In #115, we introduced the new monitoring cluster with Grafana as a
Fargate ECS service in it.

In this PR, we create the associated concourse pipeline.
This can be accessed here: https://cd.gds-reliability.engineering/teams/govuk-tools/pipelines/monitoring

The concourse pipeline works as follows:
1. deploy the base monitoring cluster
2. deploy the Grafana ECS task, i.e container.
3. configure Grafana

The task to configure Grafana is designed to:
1. wait for Grafana to be up and running, as determined by HTTP status
code 200
2. apply the Terraform Grafana config
3. handle bug https://github.com/grafana/terraform-provider-grafana/issues/129
where terraform does not recreate some grafana resources if not found

In addition, we need to add CDIO Big Concourse Egress IPs to the list of
IPs allowed to reach Grafana so that Concourse can apply against Grafana.

Future work:
1. refactor shell scripts into separate files once agreed style.

Ref:
1. [trello card](https://trello.com/c/ZJuJoV3H/350-create-monitoring-concourse-pipeline)