Promote modules/govuk to a deployment
Previously we had the govuk-test deployment, which was supposed to
configure things which were specific to the test environment and then
pass them down to the govuk module.

But then we ended up pulling (almost) all of the environment specific
configuration out into .tfvars files in variables/test/*.tfvars, so the
govuk-test deployment ended up not knowing very much about the test
environment. Instead, it just became a really annoying layer of
indirection - any variable or outputs needed to be passed through both
the deployment and the module.

Removing this layer of indirection should make our lives much easier.

That the terraform plan for this is going to be an absolute abomination,
but I think we should be able to fix that with:

```
terraform state list | while read -ra line; do
  # If the line starts with `module.govuk`
  if grep "^module\.govuk\." <<< "$line"; then
    # Then move it in the statefile to a location without the module.govuk. prefix
    terraform state mv "$line" "${line#module.govuk.}"
  fi
done
```