Change statsd namespace for ECS tasks
This will separate the metrics for apps running in ECS and EC2.

Initially we intended the metrics from both platforms to be
displayed as one platform in Grafana.

On reflection it would be better for these metrics to be
separate, and displayed in separate dashboards. This will
avoid confusion - you don't want to see metrics from one
platform while debugging another platform.

After this change, we'll use the ECS metrics in the newly
created ECS Grafana, and leave the EC2 side of things
untouched.

We may need to recreate 5-6 alerts that use app metrics.
This should be pretty trivial to do in Puppet.