Only allow ingress to frontend ALB from office IPs
This is more consistent with what was there before I refactored it to
use the public load balancer module.

In the long run, I don't think frontend needs a public load balancer -
it's just got one for now because we haven't got router running yet. So
having the office IPs passed in directly is fine for now.