Output task_network_config from publisher-web deployment
This output will make it possible to acquire both the task
to be run and the network configuration for the task in a
single terraform apply / terraform output step.

This is useful, since we can run any tasks with ECS RunTask
using the Terraform output from a single deployment.
Concretely, this will cut down on the number of tasks in
our Concourse pipeline.