Remove fetch-task-network-config Concourse task
This task is unecessary since we can export the network_config for
an ECS RunTask call directly from the Terraform statefile with
Terraform output in the app deployment.

This does tie the run-task script quite fairly tightly to applications
but I think that is OK since the interface (dumping some expected files
in a directory) could be replicated by other tasks that aren't apps.