Use ECS CLI Docker image for run-task
The ECS CLI has some additional ECS-specific tooling that
is useful for running tasks in ECS.

Particularly esc-cli logs is helpful for pulling logs related
to an ECS task and cluster from CloudWatch.

This does require us to build an image, but I don't think this
should be much of an overhead.