Create task definition JSON in the main terraform
Currently this is only wired up end-to-end for content-store. The other
apps have placeholder values, which will be filled in in future commits.

The idea here is to get terraform to template out the task definition as
JSON, which will later be passed to the AWS CLI:

```
aws ecs register-task-definition --cli-input-json file://terraform-outputs.json
```

This allows us to decouple the config management (which terraform is
good at) from the application deployments (which terraform is not good
at), and get the best of both worlds.

A few extra points of interest in this commit:

- I've prematurely moved the environment variables which I think will be
  the same in every app into a `defaults.tf` file, which declares them
  using locals. This should reduce the amount of duplication as we move
  the other apps over.
- Rather than hardcoding "eu-west-1" everywhere, I've created a
  `region.tf` file which uses the aws_region datasource.
- Rather than fetching the secrets datasources in app_content_store.tf
  I've moved them into their own file, because this feels less noisy.
- There were a few places where we needed to fetch remote state (from
  govuk-aws) - I've moved these all into `remote_state.tf`
- The new `output` for the things related to content-store is a map with
  `draft` and `live` keys, as this makes it a bit easier to pass around
  in concourse.

Once this is wired up in concourse, we'll be able to remove
deployments/apps/content-store, as that won't be needed anymore.