Use the AWS CLI not terraform to deploy content-store
The JSON needed to call `aws ecs register-task-definition --cli-input-json`
is available in the terraform outputs now.

This means that instead of needing to apply a separate terraform
deployment to create a new content-store task-definition, we can just
use the AWS CLI to create it.

To pass the terraform output from the run-terraform job to the
deploy-content-store job, we're using an S3 resource.

This means that once run-terraform completes, we pull down the terraform
outputs, and then write them to an S3 bucket. They're then downloaded,
and passed as an input to the deploy-content-store job.

The S3 resource uses versioning, so old versions of the terraform
outputs will be recoverable too.

To prevent the deploy-content-store job from triggering every time
run-terraform passes, I've done a slightly hacky thing. The
run-terraform job also takes the _old_ version of the terraform outputs
as an input, compares the new version with the old version, and if
they're the same, deletes the new version to prevent it from being
uploaded to S3.

The effect of this is that when the outputs change, a new version of the
resource is created, but otherwise it won't trigger any downstream jobs.

Other things to note in this commit:

- I've created v2 tasks, instead of trying to make the old tasks forward
  and backward compatible. We can remove the old tasks once all the apps
  are using the new ones.
- I was able to replace the govuk-infrastructure-content-store resource
  with a govuk-infrastructure-concourse-tasks resource, since the only
  things in this repo that affect the deploy-content-store job are the
  tasks in concourse/tasks now. Currently this is only used by the
  deploy-content-store job, but all the other apps should be able to use
  it too, once they're doing things in the new style.