Update deploy Concourse pipeline to deploy Publisher
This changes the Concourse pipeline so that we can deploy Publisher
in the new way - rather than having a Terraform apply create the
task definitions for Publisher, we're using the AWS CLI for this bit).

The Concourse pipeline is a bit more complex and verbose than it
would ideally be. The smoke tests, and any other task using run-task
is also broken by this commit, but those should soon be fixed.