refactor (draft-)static to use new deployment style
In PR #150, we refactored the way we deploy applications by outputting the
application task definition json when applying the govuk platforming terraform.
This json file is then diffed with previous version and if changes exist, a new
task definition is created for the app using AWS cli.

This PR refactors  (draft-)static app to use this new deployment style.

Similar PRs are #152 and #153 for other apps.

Ref:
1. [trello
card](https://trello.com/c/5JT7LE7k/382-update-the-static-application-to-use-the-new-deployment-approach)