Add IAM user, role and policy for updating ECR images
This commit adds a new user, role and policy for updating ECR images. A new `concourse_ecr_user` user is created to be able to use the `registry-image` resource in the `build-images` Concourse pipeline, which requires an AWS access key id and secret access key; these will be generated manually and set as pipeline variables, only accessible to this specific pipeline.

The `concourse_ecr_user` user will be able to assume the new `push_image_to_ecr_role` role, again specified as part of the `registry-image` resource. This role contains permissions to both read images from and write images to ECR for a limited subset of repositories.